moved list of CoxDeviance method to testing, using list of buffer version
